requirements.txt
# dbt Metadata Automation Requirements

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: Only needed if using RAG features (not required for basic metadata fetch)
# langchain>=0.1.0
# chromadb>=0.4.0
# google-generativeai>=0.3.0
